+++ src/GafferUI/BoxIONodeGadget.cpp
@@ -42,6 +42,7 @@
 #include "Gaffer/BoxIO.h"
 #include "Gaffer/BoxOut.h"
 #include "Gaffer/Metadata.h"
+#include "Gaffer/MetadataAlgo.h"
 #include "Gaffer/ScriptNode.h"
 #include "Gaffer/StringPlug.h"
 #include "Gaffer/UndoScope.h"
@@ -79,6 +80,11 @@ class BoxIOPlugAdder : public PlugAdder
 		bool canCreateConnection( const Plug *endpoint ) const override
 		{
 			if( !PlugAdder::canCreateConnection( endpoint ) )
+			{
+				return false;
+			}
+
+			if( MetadataAlgo::readOnly( m_boxIO.get() ) )
 			{
 				return false;
 			}